include zoomer in custom plotter
