+++ PlotterUtils/MyPlotPicker.hpp
@@ -1,10 +1,9 @@
-// Copyright (c) 2014-2016 Josh Blum
+// Copyright (c) 2014-2015 Josh Blum
 // SPDX-License-Identifier: BSL-1.0
 
 #pragma once
 #include "PlotUtilsConfig.hpp"
 #include <qwt_plot_zoomer.h>
-#include <QVariant>
 
 class QwtRasterData;
 
@@ -19,12 +18,6 @@ public:
 
     //! support for getting a z-axis value
     void registerRaster(QwtRasterData *raster);
-
-    //! query the zoomer state as a QVariant
-    QVariant state(void) const;
-
-    //! restore the state from QVariant
-    void setState(const QVariant &state);
 
 protected:
     QwtText trackerTextF(const QPointF &pos) const;
