+++ WaveMonitor/WaveMonitorDisplay.hpp
@@ -14,9 +14,7 @@
 class MyQwtPlot;
 class QwtPlotGrid;
 class QwtPlotCurve;
-class QwtPlotZoomer;
 class QwtPlotMarker;
-class MyPlotPicker;
 
 class WaveMonitorDisplay : public QWidget, public Pothos::Block
 {
@@ -97,14 +95,6 @@ public:
         return this->minimumSizeHint();
     }
 
-signals:
-
-    void stateChanged(const QVariant &value);
-
-public slots:
-
-    void restoreState(const QVariant &value);
-
 private slots:
     void installLegend(void);
     void handleLegendChecked(const QVariant &, bool, int);
@@ -118,7 +108,6 @@ private:
 
     MyQwtPlot *_mainPlot;
     QwtPlotGrid *_plotGrid;
-    MyPlotPicker *_zoomer;
     double _sampleRate;
     double _sampleRateWoAxisUnits;
     size_t _numPoints;
